# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BorrowBuddy - Backend Technical Test",
    "description": "This is documentation of backend technical test with borrowing book theme. In this api, i give name BorrowBuddi which implement user, book, borrow, and penalty management. This api cover all of requirement of test including some extras like RBAC, soft delete handling, and penalty management. ",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
...
  "paths": {
...
    "/users/{userId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "(ADMIN) Get specific user data",
        "description": "This can only be done by the admin.",
        "operationId": "getUserById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "useId is needed",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithBorrowedBookCountAndborrowedBooksObject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithBorrowedBookCountAndborrowedBooksObject"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "(ALL USER) update user data",
        "description": "This can only can done by all user which have login.",
        "operationId": "updateUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "useId is needed",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "(ADMIN) Delete specific user (soft delete)",
        "description": "This can only be done by admin.",
        "operationId": "deleteUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "useId is needed",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/recover": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "(ADMIN) recover user from deleted",
        "description": "This can only be done by admin to recover user which have deleted.",
        "operationId": "recoverUser",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "useId is needed",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoverUserResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/RecoverUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User account not found",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/FailedRecoverUserResponse"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/penalties/{penaltyId}": {
      "put": {
        "tags": [
          "penalty"
        ],
        "summary": "(ADMIN) update penalty",
        "description": "This can only be admin.",
        "operationId": "updatePenalty",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePenaltyRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePenaltyRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePenaltyRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "penaltyId",
            "in": "path",
            "description": "penaltyId is needed",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePenaltyUserResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePenaltyUserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "penalty"
        ],
        "summary": "(ADMIN) delete penalty",
        "description": "This can only be done by admin.",
        "operationId": "deletePenalty",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "penaltyId",
            "in": "path",
            "description": "PenaltyId is needed",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePenaltyResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePenaltyResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "code": {
            "type": "string",
            "example": "M001"
          },
          "name": {
            "type": "string",
            "example": "user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@gmail.com"
          },
          "role": {
            "type": "string",
            "example": "MEMBER"
          }
        }
      },
      "LessUserObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "code": {
            "type": "string",
            "example": "M001"
          },
          "name": {
            "type": "string",
            "example": "user"
          }
        }
      },
...
      "UserWithBorrowedBookCountAndborrowedBooksObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "code": {
            "type": "string",
            "example": "M001"
          },
          "name": {
            "type": "string",
            "example": "user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@gmail.com"
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "role": {
            "type": "string",
            "example": "MEMBER"
          },
          "borrowedBooksCount": {
            "type": "number",
            "example": 0
          },
          "borrowedBooks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookObject"
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2024-07-11T09:32:38.328Z"
          }
        }
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "user"
          },
          "email": {
            "type": "string",
            "example": "user@email.com"
          },
          "password": {
            "type": "string",
            "example": "rahasia"
          }
        }
      },
      "UpdateUserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Success update User!"
          },
          "data": {
            "$ref": "#/components/schemas/UserObject"
          }
        }
      },
      "DeleteUserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Success delete User!"
          }
        }
      },
      "RecoverUserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Success Recover User!"
          },
          "data": {
            "$ref": "#/components/schemas/UserObject"
          }
        }
      },
      "FailedRecoverUserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "User Not Found or still activated!"
          }
        }
      },
...
      "BookObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Buku"
          },
          "author": {
            "type": "string",
            "example": "Arief"
          },
          "code": {
            "type": "string",
            "example": "BOOK-1"
          },
          "stock": {
            "type": "number",
            "example": 1
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2024-07-11T09:32:38.328Z"
          }
        }
      },
...
      "PenaltyObject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "startDate": {
            "type": "string",
            "example": "2024-07-11T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2024-07-14T10:09:25.265Z"
          },
          "user": {
            "type": "object",
            "$ref": "#/components/schemas/LessUserObject"
          }
        }
      },
...
      "UpdatePenaltyRequest": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "example": "2024-07-11T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2024-07-14T00:00:00.000Z"
          }
        }
      },
      "UpdatePenaltyUserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Success update penalty !"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PenaltyObject"
            }
          }
        }
      },
      "DeletePenaltyResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Success Delete Penalty!"
          }
        }
      },
...
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}